Cleared redundant files and codeline
