+++ mainwindow.ui
@@ -29,7 +29,7 @@
        <rect>
         <x>0</x>
         <y>0</y>
-        <width>1261</width>
+        <width>931</width>
         <height>561</height>
        </rect>
       </property>
